requirements.txt
-i https://pypi.org/simple
av==11.0.0; python_version >= '3.8'
bottle==0.12.25
numpy==1.26.4; python_version >= '3.9'
picamera2==0.3.16; python_version >= '3.9'
pidng==4.0.9; python_version >= '3.6'
piexif==1.1.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pillow==10.2.0; python_version >= '3.8'
python-prctl==1.8.1
simplejpeg==1.7.2; python_version >= '3.7'
v4l2-python3==0.3.3
